refactor(clawdstrike-logos): extract verifier_core from verifier.rs
Move the PolicyVerifier struct, its Default/inherent impls (consistency,
completeness, inheritance entry points), expected-action-type derivation,
atom collection, formula classification, and the z3 counterexample helpers
into verifier/verifier_core.rs. PolicyVerifier is re-exported explicitly
(pub); the analysis free fns are widened to pub(crate) for cross-module use
via the root glob. The lone #[cfg(test)] expected_action_types_for_policy_set
helper deliberately stays in the verifier.rs parent namespace. No logic
changes; every #[cfg(feature = "z3")] attribute preserved verbatim.
